changed input to keyboard, added line render that tracks mouse
line render will be used to determine the ball trajectory eventually, i
want to give the player more control as i am already restricting vision
with the recenetly added particle clouds and its more in tune with the
spirit of my game. Keyboard movement is not working yet, the input works
fine but clamping the paddle to screen bounds is not working yet.